Elimina archivo original de recetas de máquina
