--- views/machine_recipes.py
@@ -1,1579 +1,0 @@
-# -*- coding: utf-8 -*-
-# Recetas de Máquina (layout estilo SELOGICA) — versiones + historial con grilla y export
-from .base import *  # ctk, tk, ttk, messagebox, filedialog, BASE_DIR
-import os, json, csv, sys, re
-from datetime import datetime
-from openpyxl.utils import range_boundaries, get_column_letter, column_index_from_string
-from openpyxl.cell.cell import MergedCell
-
-# --- Detectar Excel COM una sola vez ---
-try:
-    import win32com.client as win32  # pywin32
-    HAS_EXCEL_COM = True
-except Exception:
-    HAS_EXCEL_COM = False
-
-# ---------- Paths ----------
-RECIPES_DIR = os.path.join(BASE_DIR, "machine_recipes")
-HISTORY_CSV = os.path.join(RECIPES_DIR, "_history.csv")
-os.makedirs(RECIPES_DIR, exist_ok=True)
-
-DATA_SCHEMA_VERSION = 2
-
-# ---------- Excel helpers ----------
-EXCEL_MAP = {
-    # ---------- Encabezado ----------
-    "program": "DEFG:4",
-    "mould_desig": "DEFG:5",
-    "material": "DEFG:6",
-    "date_of_entry": "KLMNO:4",
-    "cavities": "KLMNO:5",
-    "machine": "KLMNO:6",
-
-    # ---------- Injection unit / Key data (superior) ----------
-    "screw_diameter_mm": "J:9",
-
-    # Key data (N–O)
-    "cycle_time_s": "NO:8",
-    "injection_time_s": "NO:9",
-    "holding_press_time_s": "NO:10",
-    "rem_cooling_time_s": "NO:11",
-    "dosage_time_s": "NO:12",
-    "screw_stroke_mm": "NO:13",
-    "mould_stroke_mm": "NO:14",
-    "ejector_stroke_mm": "NO:15",
-    "shot_weight_g": "NO:16",
-    "plasticising_flow_kg_h": "NO:17",
-    "dosage_capacity_g_s": "NO:18",
-    "dosage_volume_ccm": "NO:19",
-    "material_cushion_ccm": "NO:20",
-    "max_injection_pressure_bar": "NO:21",
-
-    # ---------- Injection (filas 11–15) ----------
-    "inj_press_limit_bar_1": "I:11",
-    "inj_press_limit_bar_2": "H:11",
-    "inj_press_limit_bar_3": "G:11",
-
-    "injection_speed_1": "I:12",
-    "injection_speed_2": "H:12",
-    "injection_speed_3": "G:12",
-
-    "end_of_stage_mm_1": "I:13",
-    "end_of_stage_mm_2": "H:13",
-    "end_of_stage_mm_3": "G:13",
-
-    "injection_flow_ccm_s_1": "I:14",
-    "injection_flow_ccm_s_2": "H:14",
-    "injection_flow_ccm_s_3": "G:14",
-
-    "end_of_stage_ccm_1": "I:15",
-    "end_of_stage_ccm_2": "H:15",
-    "end_of_stage_ccm_3": "G:15",
-
-    # ---------- Plasticizing (filas 18–20) ----------
-    "screw_speed_m_min_1": "E:18",
-    "back_pressure_bar": "E:19",
-    "plasticizing_end_stage_ccm": "E:20",
-
-    # ---------- Holding pressure (filas 23–24) ----------
-    "hold_time_s_1": "F:23",
-    "hold_time_s_2": "G:23",
-    "hold_time_s_3": "H:23",
-
-    "hold_pressure_bar_1": "F:24",
-    "hold_pressure_bar_2": "G:24",
-    "hold_pressure_bar_3": "H:24",
-
-    # ---------- Temperatures (filas 27–29) ----------
-    "cylinder_temp_c_1": "D:27",
-    "cylinder_temp_c_2": "E:27",
-    "cylinder_temp_c_3": "F:27",
-    "cylinder_temp_c_4": "G:27",
-    "cylinder_temp_c_5": "H:27",
-
-    "tolerance_c_1": "D:28",
-    "tolerance_c_2": "E:28",
-    "tolerance_c_3": "F:28",
-    "tolerance_c_4": "G:28",
-    "tolerance_c_5": "H:28",
-
-    "feed_yoke_temp_c": "D:29",
-    "lower_enable_tol_c": "J:30",
-    "upper_switch_off_tol_c": "M:31",
-
-    # ---------- Mould movements (31–33) ----------
-    # Opening
-    "opening_end_stage_mm_1": "E:31",
-    "opening_end_stage_mm_2": "F:31",
-    "opening_end_stage_mm_3": "G:31",
-    "opening_end_stage_mm_4": "H:31",
-
-    "opening_speed_mm_s_1": "E:32",
-    "opening_speed_mm_s_2": "F:32",
-    "opening_speed_mm_s_3": "G:32",
-    "opening_speed_mm_s_4": "H:32",
-
-    "opening_force_kN_1": "E:33",
-    "opening_force_kN_2": "F:33",
-    "opening_force_kN_3": "G:33",
-    "opening_force_kN_4": "H:33",
-
-    # Closing
-    "closing_end_stage_mm_1": "K:31",
-    "closing_end_stage_mm_2": "L:31",
-    "closing_end_stage_mm_3": "M:31",
-    "closing_end_stage_mm_4": "N:31",
-
-    "closing_speed_mm_s_1": "K:32",
-    "closing_speed_mm_s_2": "L:32",
-    "closing_speed_mm_s_3": "M:32",
-    "closing_speed_mm_s_4": "N:32",
-
-    "closing_force_kN_1": "K:33",
-    "closing_force_kN_2": "L:33",
-    "closing_force_kN_3": "M:33",
-    "closing_force_kN_4": "N:33",
-
-    # ---------- Clamping ----------
-    "mould_closed_kN": "D:37",
-}
-
-# (B) Traducción Excel ↔ UI (alias de claves)
-
-ALIAS_EXCEL_TO_UI = {
-    # Encabezado
-    "program": "program",
-    "mould_desig": "mould_desig",
-    "material": "material",
-    "date_of_entry": "date_of_entry",
-    "cavities": "cavities",
-    "machine": "machine",
-
-    # Injection unit / key data
-    "screw_diameter_mm": "screw_d_mm",
-
-    "cycle_time_s": "cycle_time_s",
-    "injection_time_s": "injection_time_s",
-    "holding_press_time_s": "holding_press_time_s",
-    "rem_cooling_time_s": "rem_cooling_time_s",
-    "dosage_time_s": "dosage_time_s",
-    "screw_stroke_mm": "screw_stroke_mm",
-    "mould_stroke_mm": "mould_stroke_mm",
-    "ejector_stroke_mm": "ejector_stroke_mm",
-    "shot_weight_g": "shot_weight_g",
-    "plasticising_flow_kg_h": "plasticising_flow_kgh",
-    "dosage_capacity_g_s": "dosage_capacity_gs",
-    "dosage_volume_ccm": "dosage_volume_ccm",
-    "material_cushion_ccm": "material_cushion_ccm",
-    "max_injection_pressure_bar": "max_inj_pressure_bar",
-
-    # Injection
-    "inj_press_limit_bar_1": "inj_press_lim_1",
-    "inj_press_limit_bar_2": "inj_press_lim_2",
-    "inj_press_limit_bar_3": "inj_press_lim_3",
-
-    "injection_speed_1": "inj_speed_1",
-    "injection_speed_2": "inj_speed_2",
-    "injection_speed_3": "inj_speed_3",
-
-    "end_of_stage_mm_1": "inj_end_stage_mm_1",
-    "end_of_stage_mm_2": "inj_end_stage_mm_2",
-    "end_of_stage_mm_3": "inj_end_stage_mm_3",
-
-    "injection_flow_ccm_s_1": "inj_flow_1",
-    "injection_flow_ccm_s_2": "inj_flow_2",
-    "injection_flow_ccm_s_3": "inj_flow_3",
-
-    "end_of_stage_ccm_1": "inj_end_stage_ccm_1",
-    "end_of_stage_ccm_2": "inj_end_stage_ccm_2",
-    "end_of_stage_ccm_3": "inj_end_stage_ccm_3",
-
-    # Plasticizing
-    "screw_speed_m_min_1": "plast_screw_speed",
-    "back_pressure_bar": "plast_back_pressure",
-    "plasticizing_end_stage_ccm": "plast_end_stage_ccm",
-
-    # Holding pressure
-    "hold_time_s_1": "hp_time_1",
-    "hold_time_s_2": "hp_time_2",
-    "hold_time_s_3": "hp_time_3",
-
-    "hold_pressure_bar_1": "hp_press_1",
-    "hold_pressure_bar_2": "hp_press_2",
-    "hold_pressure_bar_3": "hp_press_3",
-
-    # Temperatures
-    "cylinder_temp_c_1": "temp_c1",
-    "cylinder_temp_c_2": "temp_c2",
-    "cylinder_temp_c_3": "temp_c3",
-    "cylinder_temp_c_4": "temp_c4",
-    "cylinder_temp_c_5": "temp_c5",
-
-    "tolerance_c_1": "tol_c1",
-    "tolerance_c_2": "tol_c2",
-    "tolerance_c_3": "tol_c3",
-    "tolerance_c_4": "tol_c4",
-    "tolerance_c_5": "tol_c5",
-
-    "feed_yoke_temp_c": "feed_yoke_temp",
-    "lower_enable_tol_c": "lower_enable_tol",
-    "upper_switch_off_tol_c": "upper_switch_off_tol",
-
-    # Movements Opening
-    "opening_end_stage_mm_1": "open_end_mm_1",
-    "opening_end_stage_mm_2": "open_end_mm_2",
-    "opening_end_stage_mm_3": "open_end_mm_3",
-    "opening_end_stage_mm_4": "open_end_mm_4",
-
-    "opening_speed_mm_s_1": "open_speed_1",
-    "opening_speed_mm_s_2": "open_speed_2",
-    "opening_speed_mm_s_3": "open_speed_3",
-    "opening_speed_mm_s_4": "open_speed_4",
-
-    "opening_force_kN_1": "open_force_1",
-    "opening_force_kN_2": "open_force_2",
-    "opening_force_kN_3": "open_force_3",
-    "opening_force_kN_4": "open_force_4",
-
-    # Movements Closing
-    "closing_end_stage_mm_1": "close_end_mm_1",
-    "closing_end_stage_mm_2": "close_end_mm_2",
-    "closing_end_stage_mm_3": "close_end_mm_3",
-    "closing_end_stage_mm_4": "close_end_mm_4",
-
-    "closing_speed_mm_s_1": "close_speed_1",
-    "closing_speed_mm_s_2": "close_speed_2",
-    "closing_speed_mm_s_3": "close_speed_3",
-    "closing_speed_mm_s_4": "close_speed_4",
-
-    "closing_force_kN_1": "close_force_1",
-    "closing_force_kN_2": "close_force_2",
-    "closing_force_kN_3": "close_force_3",
-    "closing_force_kN_4": "close_force_4",
-
-    # Clamping
-    "mould_closed_kN": "mould_closed_kn",
-}
-
-NUM_FIELDS = {
-    "cycle_time_s","injection_time_s","holding_press_time_s","rem_cooling_time_s",
-    "dosage_time_s","screw_stroke_mm","mould_stroke_mm","ejector_stroke_mm",
-    "shot_weight_g","plasticising_flow_kgh","dosage_capacity_gs","dosage_volume_ccm",
-    "material_cushion_ccm","max_inj_pressure_bar","screw_d_mm","pcs_1",
-    "inj_press_lim_1","inj_press_lim_2","inj_press_lim_3",
-    "inj_speed_1","inj_speed_2","inj_speed_3",
-    "inj_end_stage_mm_1","inj_end_stage_mm_2","inj_end_stage_mm_3",
-    "inj_flow_1","inj_flow_2","inj_flow_3",
-    "inj_end_stage_ccm_1","inj_end_stage_ccm_2","inj_end_stage_ccm_3",
-    "plast_screw_speed","plast_back_pressure","plast_end_stage_ccm",
-    "hp_time_1","hp_time_2","hp_time_3",
-    "hp_press_1","hp_press_2","hp_press_3","hp_press_4",
-    "temp_c1","temp_c2","temp_c3","temp_c4","temp_c5",
-    "tol_c1","tol_c2","tol_c3","tol_c4","tol_c5",
-    "feed_yoke_temp","lower_enable_tol","upper_switch_off_tol",
-    "open_end_mm_1","open_end_mm_2","open_end_mm_3","open_end_mm_4",
-    "open_speed_1","open_speed_2","open_speed_3","open_speed_4",
-    "open_force_1","open_force_2","open_force_3","open_force_4",
-    "close_end_mm_1","close_end_mm_2","close_end_mm_3","close_end_mm_4",
-    "close_speed_1","close_speed_2","close_speed_3","close_speed_4",
-    "close_force_1","close_force_2","close_force_3","close_force_4",
-    "mould_closed_kn",
-}
-
-# === Usa tu plantilla predefinida formatorecta.xlsx ===
-# Intenta estas rutas en orden; usa la primera que exista.
-TEMPLATE_CANDIDATES = [
-    os.path.join(BASE_DIR, "formatorecta.xlsx"),
-    os.path.join(RECIPES_DIR, "formatorecta.xlsx"),
-    os.path.join(BASE_DIR, "assets", "formatorecta.xlsx"),
-    "/mnt/data/formatorecta.xlsx",  # por si ya lo tienes ahí
-]
-
-def _find_excel_template():
-    for p in TEMPLATE_CANDIDATES:
-        if os.path.exists(p):
-            return p
-    return None
-
-def _colblock_to_a1_range(spec: str) -> str:
-    """
-    'DEFG:4' -> 'D4:G4' ; 'NO:8' -> 'N8:O8' ; 'J:9' -> 'J9'
-    Si ya viene 'A1:D1' u 'A1', lo regresa igual.
-    """
-    s = spec.strip().upper()
-    if ":" not in s:
-        return s
-    left, right = s.split(":")
-    # Caso 'A1:D1'
-    if left and left[-1].isdigit():
-        return s
-    # Caso 'COLS:ROW'
-    cols, row = left, right
-    cols = "".join([c for c in cols if c.isalpha()])
-    row = "".join([c for c in row if c.isdigit()])
-    if not cols or not row:
-        return s
-    start = cols[0]
-    end = cols[-1]
-    if start == end:
-        return f"{start}{row}"
-    return f"{start}{row}:{end}{row}"
-
-def _anchor_address(ws, addr: str) -> str:
-    """
-    Devuelve la coordenada superior-izquierda válida para escribir.
-    Soporta celdas simples, rangos 'A1:D1' y nuestro formato 'DEFG:4'.
-    """
-    addr = _colblock_to_a1_range(addr)
-    # Si es rango, usa su esquina superior-izquierda
-    if ":" in addr:
-        min_col, min_row, max_col, max_row = range_boundaries(addr)
-        return f"{get_column_letter(min_col)}{min_row}"
-
-    # Si es celda individual, pero pertenece a un merge -> ancla del merge
-    cell = ws[addr]
-    if isinstance(cell, MergedCell):
-        for rng in ws.merged_cells.ranges:
-            if addr in rng:
-                min_col, min_row, *_ = range_boundaries(str(rng))
-                return f"{get_column_letter(min_col)}{min_row}"
-    return addr
-
-# Bloques del template que queremos dejar en blanco si no hay dato en UI
-CLEAR_BLOCKS = [
-    "D4:G6",   # program/mould/material (entradas)
-    "K4:O6",   # date/cavities/machine
-    "J9:J9",   # Screw Ø [mm]
-    "N8:O21",  # Key data
-    "G11:I15", # Injection table
-    "F18:F20", # Plasticizing
-    "F23:H23","F24:H24", # Holding pressure
-    "D27:H29", # Temperatures
-    "E31:H33", # Opening
-    "K31:N33", # Closing
-    "D37:D37", # Clamping
-]
-
-def _clear_blocks(ws, blocks):
-    """Limpia (pone None) sólo en la celda ancla si hay merges."""
-    seen = set()
-    for block in blocks:
-        a1 = _colblock_to_a1_range(block)
-        min_c, min_r, max_c, max_r = range_boundaries(a1)
-        for col in range(min_c, max_c + 1):
-            for row in range(min_r, max_r + 1):
-                coord = f"{get_column_letter(col)}{row}"
-                anchor = _anchor_address(ws, coord)
-                if anchor not in seen:
-                    ws[anchor].value = None
-                    seen.add(anchor)
-
-def _to_excel_value(excel_key: str, ui_key: str, raw_val):
-    s = "" if raw_val is None else str(raw_val).strip()
-    if s == "":
-        return None  # celda en blanco si no diste valor
-
-    # numérico si el campo es numérico
-    if any(k in NUM_FIELDS for k in (excel_key, ui_key)):
-        try:
-            s2 = s.replace(",", "")
-            v = float(s2)
-            return int(v) if abs(v - int(v)) < 1e-12 else v
-        except Exception:
-            pass
-    return _safe_excel(s)
-
-# --- Fallback limpio con openpyxl ---
-def _strip_links_and_drawings(wb):
-    # Quitar vínculos externos para que Excel no “repare”
-    try:
-        wb._external_links = []
-    except Exception:
-        pass
-    try:
-        rels = getattr(wb, "_rels", None)
-        if rels:
-            for rid in list(rels):
-                rel = rels[rid]
-                if "externalLink" in getattr(rel, "reltype", ""):
-                    del rels[rid]
-    except Exception:
-        pass
-
-    # Quitar drawings/legacy drawings por hoja (imágenes/autoformas)
-    for ws in wb.worksheets:
-        try:
-            ws._images = []
-        except Exception:
-            pass
-        try:
-            ws._charts = []
-        except Exception:
-            pass
-        try:
-            ws._drawing = None
-        except Exception:
-            pass
-        try:
-            if hasattr(ws, "_rels"):
-                for rid in list(ws._rels):
-                    if "drawing" in getattr(ws._rels[rid], "reltype", ""):
-                        del ws._rels[rid]
-        except Exception:
-            pass
-        try:
-            ws.legacy_drawing = None
-        except Exception:
-            pass
-        # Ocultar ceros en la vista
-        try:
-            ws.sheet_view.showZeros = False
-        except Exception:
-            pass
-
-def _purge_all_formulas(wb):
-    """Elimina TODA fórmula del libro (incluidas compartidas) para que Excel no 'repare'."""
-    for ws in wb.worksheets:
-        # romper metadatos de fórmulas compartidas
-        try:
-            ws.formula_attributes = {}
-        except Exception:
-            pass
-
-        # recorrer rango usado y borrar fórmulas
-        dim = ws.calculate_dimension()
-        if not dim or ":" not in dim:
-            continue
-        min_c, min_r, max_c, max_r = range_boundaries(dim)
-        for row in ws.iter_rows(min_row=min_r, max_row=max_r, min_col=min_c, max_col=max_c):
-            for cell in row:
-                if isinstance(cell, MergedCell):
-                    continue
-                try:
-                    if getattr(cell, "data_type", None) == "f" or (isinstance(cell.value, str) and cell.value.startswith("=")):
-                        cell.value = None
-                except Exception:
-                    pass
-
-    # algunos nombres definidos son fórmulas -> quitarlos
-    try:
-        keep = []
-        for dn in list(wb.defined_names.definedName):
-            if getattr(dn, "attr_text", "") and str(dn.attr_text).startswith("="):
-                continue
-            keep.append(dn)
-        wb.defined_names.definedName = keep
-    except Exception:
-        pass
-
-def _export_snapshot_to_template(snapshot: dict, out_path: str, template_path: str, sheet_name: str | None = None):
-    from openpyxl import load_workbook
-    keep_vba = template_path.lower().endswith((".xlsm", ".xlsb"))
-    wb = load_workbook(template_path, data_only=False, keep_vba=keep_vba, keep_links=False)
-    ws = wb[sheet_name] if (sheet_name and sheet_name in wb.sheetnames) else wb.active
-
-    # Limpiar bloques que mapeas
-    _clear_blocks(ws, CLEAR_BLOCKS)
-
-    # Escribir solo campos con valor
-    for excel_key, spec in EXCEL_MAP.items():
-        ui_key = ALIAS_EXCEL_TO_UI.get(excel_key, excel_key)
-        val = snapshot.get(ui_key, "")
-        target = _anchor_address(ws, spec)
-        ws[target].value = _to_excel_value(excel_key, ui_key, val)
-
-    # Eliminar externalLinks/drawings para que Excel NO "repairs"
-    _strip_links_and_drawings(wb)
-
-    # eliminar TODAS las fórmulas del libro (evita el “se han quitado fórmulas”)
-    _purge_all_formulas(wb)
-
-    wb.save(out_path)
-
-def _a1_from_spec(spec: str) -> str:
-    """Usa el parser "COLS:ROW" y toma la esquina superior-izquierda."""
-    a1 = _colblock_to_a1_range(spec)
-    return a1.split(":")[0] if ":" in a1 else a1
-
-def _normalize_win_path(p: str) -> str:
-    """Normaliza rutas en Windows: absolutas, con barras '\\' y sin espacios/puntos finales."""
-    p = os.path.abspath(p).replace("/", "\\")
-    return p.rstrip(" .")
-
-def _export_with_excel_com(snapshot: dict, out_path: str, template_path: str, sheet_name: str | None = None):
-    import tempfile, shutil, uuid
-    import pywintypes  # viene con pywin32
-
-    # --- normaliza y asegura carpeta destino ---
-    out_path = _normalize_win_path(out_path)
-    dest_dir = os.path.dirname(out_path)
-    os.makedirs(dest_dir, exist_ok=True)
-
-    excel = win32.gencache.EnsureDispatch("Excel.Application")
-    excel.Visible = False
-    excel.DisplayAlerts = False
-    excel.AskToUpdateLinks = False
-
-    wb = excel.Workbooks.Open(template_path, UpdateLinks=0, ReadOnly=False, CorruptLoad=0)
-    try:
-        excel.Calculation = -4135  # xlCalculationManual
-        wb.CheckCompatibility = False
-
-        ws = wb.Worksheets(sheet_name) if sheet_name else wb.Worksheets(1)
-
-        for excel_key, spec in EXCEL_MAP.items():
-            ui_key = ALIAS_EXCEL_TO_UI.get(excel_key, excel_key)
-            raw = snapshot.get(ui_key, "")
-            val = "" if raw is None or str(raw).strip() == "" else raw
-            addr = _a1_from_spec(spec)
-            ws.Range(addr).Value = val
-
-        # --- guardado robusto: a %TEMP% y luego mover ---
-        ext = ".xlsm" if out_path.lower().endswith((".xlsm", ".xlsb")) else ".xlsx"
-        fmt = 52 if ext == ".xlsm" else 51
-        tmp_path = os.path.join(tempfile.gettempdir(), f"~export_{uuid.uuid4().hex}{ext}")
-
-        wb.SaveAs(Filename=tmp_path, FileFormat=fmt, ConflictResolution=2)
-
-    except pywintypes.com_error as e:
-        try:
-            wb.Close(SaveChanges=False)
-        except Exception:
-            pass
-        excel.Quit()
-        raise
-
-    wb.Close(SaveChanges=False)
-    excel.Quit()
-
-    try:
-        if os.path.exists(out_path):
-            os.remove(out_path)
-    except PermissionError:
-        # si está bloqueado, intenta sobreescritura atómica
-        pass
-    shutil.move(tmp_path, out_path)
-
-# ---------- FS helpers ----------
-def _safe_id(s: str) -> str:
-    return str(s).replace("/", "_").replace("\\", "_").strip()
-
-def _path_json(mold_id: str) -> str:
-    return os.path.join(RECIPES_DIR, f"{_safe_id(mold_id)}.json")
-
-def _versions_dir(mold_id: str) -> str:
-    d = os.path.join(RECIPES_DIR, _safe_id(mold_id), "_versions")
-    os.makedirs(d, exist_ok=True)
-    return d
-
-def _list_versions(mold_id: str):
-    d = _versions_dir(mold_id)
-    items = []
-    rx = re.compile(r"^v(\d{3})\.json$")
-    for fname in os.listdir(d):
-        m = rx.match(fname)
-        if m:
-            items.append((m.group(0).replace(".json",""), os.path.join(d, fname)))
-    items.sort(key=lambda t: t[0])
-    return items
-
-def _next_version_name(mold_id: str) -> str:
-    items = _list_versions(mold_id)
-    if not items: return "v001"
-    last = items[-1][0]
-    return f"v{int(last[1:])+1:03d}"
-
-def _load_json(mold_id: str) -> dict:
-    p = _path_json(mold_id)
-    if os.path.exists(p):
-        try:
-            with open(p, "r", encoding="utf-8") as f:
-                return json.load(f)
-        except Exception:
-            pass
-    return {}
-
-def _save_json(mold_id: str, data: dict):
-    payload = dict(data)
-    payload["_meta"] = {
-        "schema": DATA_SCHEMA_VERSION,
-        "updated_at": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
-    }
-    with open(_path_json(mold_id), "w", encoding="utf-8") as f:
-        json.dump(payload, f, ensure_ascii=False, indent=2)
-
-def _save_version_snapshot(mold_id: str, data: dict, usuario: str, motivo: str, diffs_text: str) -> str:
-    ver = _next_version_name(mold_id)
-    snap = dict(data)
-    snap["_meta"] = {
-        "schema": DATA_SCHEMA_VERSION,
-        "saved_at": datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
-        "usuario": usuario,
-        "motivo": motivo,
-        "version": ver,
-        "diffs": diffs_text,
-    }
-    out = os.path.join(_versions_dir(mold_id), f"{ver}.json")
-    with open(out, "w", encoding="utf-8") as f:
-        json.dump(snap, f, ensure_ascii=False, indent=2)
-    return ver
-
-def _load_version_snapshot(mold_id: str, version: str) -> dict:
-    path = os.path.join(_versions_dir(mold_id), f"{version}.json")
-    if not os.path.exists(path): return {}
-    with open(path, "r", encoding="utf-8") as f:
-        return json.load(f)
-
-def _cast_numeric(val: str) -> str:
-    if val is None: return ""
-    s = str(val).strip().replace(",", ".")
-    if s == "": return ""
-    try:
-        num = float(s)
-        return str(int(num)) if abs(num - int(num)) < 1e-9 else f"{num:.3f}".rstrip("0").rstrip(".")
-    except Exception:
-        return str(val).strip()
-
-def _ascii(s: str) -> str:
-    """Evita errores latin-1 al exportar PDFs en algunos backends."""
-    if s is None: return ""
-    return (str(s).replace("\u2014","-").encode("latin-1","replace").decode("latin-1"))
-
-# ---------- Sanitizers para export ----------
-try:
-    from openpyxl.cell.cell import ILLEGAL_CHARACTERS_RE
-except Exception:
-    ILLEGAL_CHARACTERS_RE = None
-
-def _safe_excel(val: str) -> str:
-    s = "" if val is None else str(val)
-    if ILLEGAL_CHARACTERS_RE:
-        s = ILLEGAL_CHARACTERS_RE.sub("", s)
-    return s
-
-def _safe_pdf(val: str) -> str:
-    # Latin-1 friendly + sin saltos raros
-    return _ascii(val).replace("\r", " ").replace("\n", " ")
-
-# ============================================================
-#                          VIEW
-# ============================================================
-class MachineRecipesView(ctk.CTkFrame):
-    """Panel de recetas de máquina con layout SELOGICA + historial/versiones."""
-
-    def __init__(self, master, app):
-        super().__init__(master, fg_color="transparent")
-        self.app = app
-        self.vars = {}          # id -> StringVar
-        self.current_mold = ""
-        self._dirty = False
-        self._preview_version = None
-        self._status_var = tk.StringVar(value="")
-        self._motivo_ph = (
-            "Formato sugerido: [Qué cambiaste] -> [Por qué] -> [Resultado esperado].\n"
-            "Ej.: Aumenté HP de 60->70 bar para mejorar compacción; espero eliminar porosidad en esquina A."
-        )
-        self._build()
-
-    # --------------------------- UI ---------------------------
-    def _build(self):
-        # Header
-        header = ctk.CTkFrame(self, corner_radius=0, fg_color=("white", "#0e1117"))
-        header.pack(fill="x", side="top")
-        left = ctk.CTkFrame(header, fg_color="transparent"); left.pack(side="left", padx=16, pady=10)
-        ctk.CTkButton(left, text="← Menú", width=110, corner_radius=10,
-                      fg_color="#E5E7EB", text_color="#111", hover_color="#D1D5DB",
-                      command=self._back_menu_guard).pack(side="left", padx=(0, 10))
-        title = ctk.CTkFrame(left, fg_color="transparent"); title.pack(side="left")
-        ctk.CTkLabel(title, text="Recetas de Máquina", font=ctk.CTkFont("Helvetica", 22, "bold")).pack(anchor="w")
-        ctk.CTkLabel(title, text="Layout estilo SELOGICA • Historial y versiones",
-                     text_color=("#6b7280", "#9CA3AF")).pack(anchor="w")
-
-        # Toolbar
-        tools = ctk.CTkFrame(self, fg_color="transparent"); tools.pack(fill="x", padx=16, pady=(8, 0))
-        opciones = ["— Selecciona —"] + list(getattr(self.app, "recipe_map", {}).keys())
-        self.mold_var = tk.StringVar(value=opciones[0])
-        self.mold_menu = ctk.CTkOptionMenu(tools, values=opciones, variable=self.mold_var, width=260,
-                                           command=lambda *_: self._on_pick_mold())
-        self.mold_menu.pack(side="left", padx=(0, 8))
-        ctk.CTkButton(tools, text="📥 Importar Excel", command=self._import_excel).pack(side="left", padx=6)
-        ctk.CTkButton(tools, text="🕓 Historial / Versiones", command=self._open_history).pack(side="left", padx=6)
-        ctk.CTkButton(tools, text="📂 Carpeta de recetas",
-                      command=lambda: self._open_folder(RECIPES_DIR)).pack(side="left", padx=6)
-
-        # === DOCUMENTAR CAMBIO ===
-        doc_card = ctk.CTkFrame(self, corner_radius=12, fg_color=("white", "#111827"))
-        doc_card.pack(fill="x", padx=16, pady=(10, 0))
-        topbar = ctk.CTkFrame(doc_card, fg_color="transparent"); topbar.pack(fill="x", padx=12, pady=(10, 4))
-        ctk.CTkLabel(topbar, text="Documentar cambio", font=ctk.CTkFont(size=14, weight="bold")).pack(side="left")
-        actions = ctk.CTkFrame(topbar, fg_color="transparent"); actions.pack(side="right")
-        ctk.CTkButton(actions, text="💾 Guardar (Ctrl+S)", command=self._save).pack(side="left")
-        ctk.CTkButton(actions, text="🗑", fg_color="#ef4444", hover_color="#dc2626", width=44,
-                      command=self._clear_all).pack(side="left", padx=(8,0))
-        body = ctk.CTkFrame(doc_card, fg_color="transparent"); body.pack(fill="both", padx=12, pady=(0, 12))
-        self.motivo_txt = ctk.CTkTextbox(body, height=110)
-        self.motivo_txt.pack(fill="x", expand=False)
-        self._init_motivo_placeholder()
-        btns = ctk.CTkFrame(body, fg_color="transparent"); btns.pack(fill="x", pady=(6,0))
-        ctk.CTkButton(btns, text="Ejemplo", width=80, command=self._insert_motivo_example).pack(side="left")
-
-        # status
-        ctk.CTkLabel(self, textvariable=self._status_var,
-                     text_color=("#6b7280", "#9CA3AF"), anchor="w").pack(fill="x", padx=16, pady=(4,0))
-
-        # CONTENIDO con scroll (formulario)
-        self.scroll = ctk.CTkScrollableFrame(self, fg_color="transparent")
-        self.scroll.pack(fill="both", expand=True, padx=16, pady=16)
-
-        sheet = ctk.CTkFrame(self.scroll, corner_radius=14, fg_color=("white", "#111827"))
-        sheet.pack(fill="x", padx=2, pady=2)
-        sheet.grid_columnconfigure(0, weight=3)
-        sheet.grid_columnconfigure(1, weight=2)
-
-        hdr = self._card(sheet, "Parameter overview (Cabecera)", row=0, col=0)
-        self._header_grid(hdr)
-        key = self._card(sheet, "Key data", row=0, col=1)
-        self._keydata_table(key)
-        blocks = self._card(sheet, "", row=1, col=0, col_span=2)
-        self._full_left_side(blocks)
-
-        self._register_shortcuts_once()
-
-    # ----- small UI builders -----
-    def _card(self, parent, title, row=0, col=0, col_span=1):
-        card = ctk.CTkFrame(parent, corner_radius=14, fg_color=("white", "#111827"))
-        card.grid(row=row, column=col, columnspan=col_span, padx=8, pady=8, sticky="nsew")
-        if title:
-            ctk.CTkLabel(card, text=title, font=ctk.CTkFont("Helvetica", 13, "bold")).pack(anchor="w", padx=10, pady=(10, 0))
-            ctk.CTkFrame(card, height=1, fg_color=("#E5E7EB", "#2B2B2B")).pack(fill="x", padx=10, pady=(6, 10))
-        return card
-
-    def _lbl(self, parent, text, **grid):
-        l = ctk.CTkLabel(parent, text=text); l.grid(**grid); return l
-
-    def _ent(self, parent, fid, w=110, justify="center", **grid):
-        v = self.vars.setdefault(fid, tk.StringVar())
-        e = ctk.CTkEntry(parent, textvariable=v, width=w, justify=justify)
-        e.grid(**grid)
-        v.trace_add("write", lambda *_: self._mark_dirty())
-        return e
-
-    def _header_grid(self, parent):
-        g = ctk.CTkFrame(parent, fg_color="transparent"); g.pack(fill="x", padx=10, pady=10)
-        for i in range(6): g.grid_columnconfigure(i, weight=1)
-        self._lbl(g, "Program",        row=0, column=0, sticky="w", padx=4, pady=4)
-        self._ent(g, "program",        row=0, column=1, sticky="ew", padx=4, pady=4)
-        self._lbl(g, "Date of entry:", row=0, column=2, sticky="w", padx=4, pady=4)
-        self._ent(g, "date_of_entry",  row=0, column=3, sticky="ew", padx=4, pady=4)
-        self._lbl(g, "Cavities",       row=0, column=4, sticky="w", padx=4, pady=4)
-        self._ent(g, "cavities",       row=0, column=5, sticky="ew", padx=4, pady=4)
-        self._lbl(g, "Mould desig.",   row=1, column=0, sticky="w", padx=4, pady=4)
-        self._ent(g, "mould_desig",    row=1, column=1, sticky="ew", padx=4, pady=4)
-        self._lbl(g, "Machine",        row=1, column=2, sticky="w", padx=4, pady=4)
-        self._ent(g, "machine",        row=1, column=3, sticky="ew", padx=4, pady=4)
-        self._lbl(g, "Material",       row=2, column=0, sticky="w", padx=4, pady=4)
-        self._ent(g, "material",       row=2, column=1, sticky="ew", padx=4, pady=4)
-
-    def _keydata_table(self, parent):
-        grid = ctk.CTkFrame(parent, fg_color="transparent"); grid.pack(fill="x", padx=10, pady=10)
-        for i in range(2): grid.grid_columnconfigure(i, weight=1)
-        rows = [
-            ("cycle_time_s",           "Cycle time [s]"),
-            ("injection_time_s",       "Injection time [s]"),
-            ("holding_press_time_s",   "Holding press. time [s]"),
-            ("rem_cooling_time_s",     "Rem. cooling time [s]"),
-            ("dosage_time_s",          "Dosage time [s]"),
-            ("screw_stroke_mm",        "Screw stroke [mm]"),
-            ("mould_stroke_mm",        "Mould stroke [mm]"),
-            ("ejector_stroke_mm",      "Ejector stroke [mm]"),
-            ("shot_weight_g",          "Shot weight [g]"),
-            ("plasticising_flow_kgh",  "Plasticising flow [kg/h]"),
-            ("dosage_capacity_gs",     "Dosage capacity [g/s]"),
-            ("dosage_volume_ccm",      "Dosage volume [ccm]"),
-            ("material_cushion_ccm",   "Material cushion [ccm]"),
-            ("max_inj_pressure_bar",   "max. inj. pressure [bar]"),
-        ]
-        for r, (fid, label) in enumerate(rows):
-            self._lbl(grid, label, row=r, column=0, sticky="w", padx=4, pady=4)
-            self._ent(grid, fid, row=r, column=1, sticky="ew", padx=4, pady=4)
-
-    def _full_left_side(self, parent):
-        ctk.CTkLabel(parent, text="Injection unit", font=ctk.CTkFont("Helvetica", 12, "bold")
-                     ).pack(anchor="center", pady=(4, 0))
-        inj_unit = ctk.CTkFrame(parent, fg_color="transparent"); inj_unit.pack(fill="x", padx=10, pady=(0, 8))
-        inj_unit.grid_columnconfigure((0,1,2,3,4), weight=1)
-        self._lbl(inj_unit, "Screw Ø [mm]", row=0, column=0, sticky="e", padx=6, pady=4)
-        self._ent(inj_unit, "screw_d_mm", w=80, row=0, column=1, sticky="w", padx=4, pady=4)
-        self._lbl(inj_unit, "Pcs. 1", row=0, column=3, sticky="e", padx=6, pady=4)
-        self._ent(inj_unit, "pcs_1", w=80, row=0, column=4, sticky="w", padx=4, pady=4)
-
-        self._section_table(parent, "Injection", [
-            ("Injection press. limiting [bar]", ["inj_press_lim_1", "inj_press_lim_2", "inj_press_lim_3"]),
-            ("Injection speed [mm/s]",          ["inj_speed_1", "inj_speed_2", "inj_speed_3"]),
-            ("End of stage [mm]",               ["inj_end_stage_mm_1", "inj_end_stage_mm_2", "inj_end_stage_mm_3"]),
-            ("Injection flow [ccm/s]",          ["inj_flow_1", "inj_flow_2", "inj_flow_3"]),
-            ("End of stage [ccm]",              ["inj_end_stage_ccm_1", "inj_end_stage_ccm_2", "inj_end_stage_ccm_3"]),
-        ])
-
-        self._section_table(parent, "Plasticizing (St.1)", [
-            ("Screw speed [m/min]",             ["plast_screw_speed"]),
-            ("Back pressure [bar]",             ["plast_back_pressure"]),
-            ("End of stage [ccm]",              ["plast_end_stage_ccm"]),
-        ])
-
-        self._section_table(parent, "Holding pressure (Pcs.2)", [
-            ("Time [s]",                        ["hp_time_1", "hp_time_2", "hp_time_3"]),
-            ("Pressure [bar]",                  ["hp_press_1", "hp_press_2", "hp_press_3", "hp_press_4"]),
-        ])
-
-        self._section_table(parent, "Temperatures (1..5)", [
-            ("Cylinder temp. [°C]",             ["temp_c1", "temp_c2", "temp_c3", "temp_c4", "temp_c5"]),
-            ("Tolerances [°C]",                 ["tol_c1", "tol_c2", "tol_c3", "tol_c4", "tol_c5"]),
-            ("Feed yoke temperature [°C]",      ["feed_yoke_temp"]),
-            ("Lower enable tol. [°C]",          ["lower_enable_tol"]),
-            ("Upper switch-off tol. [°C]",      ["upper_switch_off_tol"]),
-        ])
-
-        self._section_table(parent, "Mould movements — Opening (St.1 / St.2 / St.3)", [
-            ("End of stage [mm]",               ["open_end_mm_1", "open_end_mm_2", "open_end_mm_3"]),
-            ("Speed [mm/s]",                    ["open_speed_1", "open_speed_2", "open_speed_3"]),
-            ("Force [kN]",                      ["open_force_1", "open_force_2", "open_force_3"]),
-        ])
-
-        self._section_table(parent, "Mould movements — Closing (St.1 / St.2 / St.3 / An. HD)", [
-            ("End of stage [mm]",               ["close_end_mm_1", "close_end_mm_2", "close_end_mm_3", "close_end_mm_4"]),
-            ("Speed [mm/s]",                    ["close_speed_1", "close_speed_2", "close_speed_3", "close_speed_4"]),
-            ("Force [kN]",                      ["close_force_1", "close_force_2", "close_force_3"]),
-        ])
-
-        self._section_table(parent, "Clamping", [
-            ("Mould closed [kN]",               ["mould_closed_kn"]),
-        ])
-
-    def _section_table(self, parent, title, rows):
-        card = ctk.CTkFrame(parent, corner_radius=12, fg_color=("white", "#111a27"))
-        card.pack(fill="x", padx=10, pady=8)
-        ctk.CTkLabel(card, text=title, font=ctk.CTkFont("Helvetica", 12, "bold")).pack(anchor="w", padx=10, pady=(8, 6))
-        grid = ctk.CTkFrame(card, fg_color="transparent"); grid.pack(fill="x", padx=10, pady=(0,10))
-        max_cols = max(len(cols) for _, cols in rows)
-        for i in range(max_cols + 1): grid.grid_columnconfigure(i, weight=1)
-        for r, (label, ids) in enumerate(rows):
-            self._lbl(grid, label, row=r, column=0, sticky="w", padx=4, pady=4)
-            for c, fid in enumerate(ids, start=1):
-                self._ent(grid, fid, w=90, row=r, column=c, sticky="ew", padx=4, pady=4)
-
-    # --------------------------- Eventos básicos ---------------------------
-    def _register_shortcuts_once(self):
-        root = self.winfo_toplevel()
-        if getattr(root, "_recipes_shortcuts_ok", False): return
-        root.bind("<Control-s>", lambda e: (self._save(), "break"), add="+")
-        root._recipes_shortcuts_ok = True
-
-    def _init_motivo_placeholder(self):
-        # placeholder gris
-        self._motivo_placeholder_on = True
-        self.motivo_txt.delete("1.0", "end")
-        self.motivo_txt.insert("1.0", self._motivo_ph)
-        self.motivo_txt.configure(text_color=("#9CA3AF", "#6b7280"))
-        self.motivo_txt.bind("<FocusIn>", self._motivo_focus_in)
-        self.motivo_txt.bind("<FocusOut>", self._motivo_focus_out)
-
-    def _motivo_focus_in(self, *_):
-        if self._motivo_placeholder_on:
-            self.motivo_txt.delete("1.0", "end")
-            self.motivo_txt.configure(text_color=("black", "white"))
-            self._motivo_placeholder_on = False
-
-    def _motivo_focus_out(self, *_):
-        content = (self.motivo_txt.get("1.0", "end").strip())
-        if not content:
-            self._init_motivo_placeholder()
-
-    def _get_motivo_text(self) -> str:
-        if self._motivo_placeholder_on: return ""
-        return self.motivo_txt.get("1.0", "end").strip()
-
-    def _set_motivo_text(self, text: str):
-        self._motivo_placeholder_on = False
-        self.motivo_txt.delete("1.0", "end")
-        self.motivo_txt.insert("1.0", text)
-        self.motivo_txt.configure(text_color=("black", "white"))
-
-    def _insert_motivo_example(self):
-        example = ("Ajusté velocidad de inyección 50->65 mm/s para mejorar llenado en nervaduras; "
-                   "espero reducir cortinas y estabilizar peso del disparo (+0.2g).")
-        cur = self._get_motivo_text()
-        if not cur:
-            self._set_motivo_text(example)
-        else:
-            if not cur.endswith("\n"):
-                self.motivo_txt.insert("end", "\n")
-            self.motivo_txt.insert("end", example)
-
-    def _mark_dirty(self): self._dirty = True
-
-    def _back_menu_guard(self):
-        if self._dirty and not messagebox.askyesno("Cambios sin guardar", "Tienes cambios sin guardar. ¿Descartar?"):
-            return
-        self._dirty = False
-        self._preview_version = None
-        self._status_var.set("")
-        if hasattr(self.app, "go_menu"):
-            self.app.go_menu()
-
-    def _on_pick_mold(self):
-        mid = (self.mold_var.get() or "").strip()
-        self.current_mold = "" if (not mid or mid.startswith("—")) else mid
-        for v in self.vars.values(): v.set("")
-        self._dirty = False
-        self._preview_version = None
-        self._status_var.set("")
-        self._init_motivo_placeholder()
-        if self.current_mold:
-            data = _load_json(self.current_mold)
-            for k, v in self.vars.items():
-                v.set(str(data.get(k, "")))
-            if "date_of_entry" in self.vars and not self.vars["date_of_entry"].get().strip():
-                self.vars["date_of_entry"].set(datetime.now().strftime("%Y-%m-%d"))
-
-    def _collect(self) -> dict:
-        out = {}
-        for k, v in self.vars.items():
-            val = (v.get() or "").strip()
-            if k in NUM_FIELDS: val = _cast_numeric(val)
-            out[k] = val
-        return out
-
-    def _resolve_usuario(self) -> str:
-        try:
-            op = getattr(self.app, "operador", "")
-            return op.get() if hasattr(op, "get") else str(op or "")
-        except Exception:
-            return ""
-
-    def _save(self, *_):
-        if not self.current_mold:
-            messagebox.showwarning("Molde", "Selecciona un molde."); return
-        motivo = self._get_motivo_text()
-        if not motivo:
-            messagebox.showwarning("Motivo requerido", "Escribe el motivo del cambio para guardar."); return
-        if "date_of_entry" in self.vars and not self.vars["date_of_entry"].get().strip():
-            self.vars["date_of_entry"].set(datetime.now().strftime("%Y-%m-%d"))
-
-        old = _load_json(self.current_mold)
-        new = self._collect()
-
-        diffs = []
-        for k, nv in new.items():
-            ov = str(old.get(k, ""))
-            if (nv or "") != (ov or ""):
-                diffs.append(f"{k}: '{ov}' -> '{nv}'")
-        if not diffs:
-            messagebox.showinfo("Recetas de Máquina", "No hubo cambios en los parámetros.")
-            return
-
-        _save_json(self.current_mold, new)
-
-        usuario = self._resolve_usuario()
-        diffs_text = " | ".join(diffs)
-        ver = _save_version_snapshot(self.current_mold, new, usuario, motivo, diffs_text)
-
-        exists = os.path.exists(HISTORY_CSV)
-        with open(HISTORY_CSV, "a", newline="", encoding="utf-8") as f:
-            w = csv.writer(f)
-            if not exists:
-                w.writerow(["ts", "molde_id", "usuario", "motivo", "cambios", "version"])
-            w.writerow([datetime.now().strftime("%Y-%m-%d %H:%M:%S"),
-                        self.current_mold, usuario, motivo, diffs_text, ver])
-
-        self._init_motivo_placeholder()
-        self._dirty = False
-        self._preview_version = None
-        self._status_var.set(f"Guardado como {ver}.")
-        messagebox.showinfo("Recetas de Máquina", f"Parámetros guardados.\nSe creó snapshot: {ver}")
-
-    def _clear_all(self):
-        if not self.current_mold:
-            if messagebox.askyesno("Limpiar", "¿Limpiar todos los campos visibles?"):
-                for v in self.vars.values(): v.set("")
-            return
-        if messagebox.askyesno("Limpiar", f"¿Limpiar todos los campos del molde {self.current_mold}?"):
-            for v in self.vars.values(): v.set("")
-            self._dirty = True
-
-    # --------------------------- Historial / Versiones ---------------------------
-    def _open_history(self):
-        if not self.current_mold:
-            messagebox.showwarning("Historial", "Selecciona un molde para ver su historial."); return
-
-        top = tk.Toplevel(self); top.title(f"Historial y versiones — {self.current_mold}")
-        try: top.state('zoomed')
-        except Exception: top.geometry("1280x800+60+40")
-
-        wrap = ctk.CTkFrame(top, fg_color=("white", "#0e1117")); wrap.pack(fill="both", expand=True)
-
-        # Barra superior
-        bar = ctk.CTkFrame(wrap, fg_color="transparent"); bar.pack(fill="x", padx=12, pady=(10, 6))
-        ctk.CTkLabel(bar, text=f"Molde: {self.current_mold}", font=ctk.CTkFont(size=16, weight="bold")).pack(side="left")
-        srch_var = tk.StringVar()
-        ctk.CTkEntry(bar, textvariable=srch_var, placeholder_text="Buscar en motivo/cambios...", width=380
-                    ).pack(side="left", padx=(12,4))
-        ctk.CTkButton(bar, text="Buscar", command=lambda: load_versions()).pack(side="left")
-        ctk.CTkButton(bar, text="Mostrar todo", command=lambda: (srch_var.set(""), load_versions())
-                     ).pack(side="left", padx=(8,0))
-        ctk.CTkButton(bar, text="Abrir carpeta de versiones",
-                      command=lambda: self._open_folder(_versions_dir(self.current_mold))
-                     ).pack(side="right")
-        ctk.CTkButton(bar, text="Cerrar", fg_color="#6b7280", hover_color="#4b5563",
-                      command=top.destroy).pack(side="right", padx=(8,0))
-
-        # Paneles
-        main = ctk.CTkFrame(wrap, fg_color="transparent"); main.pack(fill="both", expand=True, padx=12, pady=(0,12))
-        main.grid_columnconfigure(0, weight=3)
-        main.grid_columnconfigure(1, weight=5)
-        main.grid_rowconfigure(0, weight=1)
-
-        # IZQUIERDA: versiones
-        left = ctk.CTkFrame(main, fg_color=("white", "#111827")); left.grid(row=0, column=0, sticky="nsew", padx=(0,8))
-        cols = ("version","fecha","usuario","motivo","cambios")
-        tree = ttk.Treeview(left, columns=cols, show="headings", height=24)
-        headers = [("version","Versión",80),("fecha","Fecha/Hora",150),("usuario","Usuario",140),
-                   ("motivo","Motivo (resumen)",280),("cambios","Cambios (preview)",400)]
-        for key, text, w in headers:
-            tree.heading(key, text=text); tree.column(key, width=w, anchor="w")
-        vsb = ttk.Scrollbar(left, orient="vertical", command=tree.yview); tree.configure(yscrollcommand=vsb.set)
-        tree.pack(side="left", fill="both", expand=True, padx=12, pady=12); vsb.pack(side="left", fill="y", pady=12)
-
-        # DERECHA: detalle (grilla)
-        right = ctk.CTkFrame(main, fg_color=("white", "#111827")); right.grid(row=0, column=1, sticky="nsew", padx=(8,0))
-        header_box = ctk.CTkFrame(right, fg_color="transparent"); header_box.pack(fill="x", padx=12, pady=(12,0))
-        detail_title = ctk.CTkLabel(header_box, text="Detalle de versión", font=ctk.CTkFont(size=14, weight="bold"))
-        detail_title.pack(side="left")
-
-        export_box = ctk.CTkFrame(header_box, fg_color="transparent"); export_box.pack(side="right")
-        btn_xlsx = ctk.CTkButton(export_box, text="⬇ Exportar Excel", state="disabled")
-        btn_pdf  = ctk.CTkButton(export_box, text="⬇ Exportar PDF", state="disabled")
-        btn_pdf.pack(side="right", padx=(8,0)); btn_xlsx.pack(side="right")
-
-        detail_wrap = ctk.CTkScrollableFrame(right, fg_color=("white","#111827"))
-        detail_wrap.pack(fill="both", expand=True, padx=12, pady=(6,12))
-
-        # placeholder de secciones
-        def clear_detail():
-            for w in detail_wrap.winfo_children():
-                w.destroy()
-
-        # ----- grilla estilo Excel por secciones -----
-        def section(title):
-            card = ctk.CTkFrame(detail_wrap, corner_radius=12, fg_color=("white", "#111a27"))
-            card.pack(fill="x", padx=6, pady=6)
-            ctk.CTkLabel(card, text=title, font=ctk.CTkFont("Helvetica", 12, "bold")).pack(anchor="w", padx=10, pady=(10,6))
-            grid = ctk.CTkFrame(card, fg_color="transparent"); grid.pack(fill="x", padx=10, pady=(0,10))
-            return grid
-
-        def row_labeled(grid, r, label, values, start_col=0):
-            ctk.CTkLabel(grid, text=label).grid(row=r, column=start_col, sticky="w", padx=4, pady=4)
-            for i, val in enumerate(values, start=1):
-                e = ctk.CTkEntry(grid, width=110, justify="center")
-                e.insert(0, str(val))
-                e.configure(state="disabled")
-                e.grid(row=r, column=start_col+i, sticky="ew", padx=4, pady=4)
-
-        # datos actuales en detalle
-        last_snap = {"_meta": {}}
-
-        def load_versions():
-            tree.delete(*tree.get_children())
-            q = (srch_var.get() or "").lower()
-            snaps = _list_versions(self.current_mold)
-            if not snaps:
-                current = _load_json(self.current_mold)
-                if current:
-                    usuario = self._resolve_usuario()
-                    motivo = "(baseline) Estado inicial importado desde receta actual"
-                    _save_version_snapshot(self.current_mold, current, usuario, motivo, "")
-                    snaps = _list_versions(self.current_mold)
-            for ver, path in snaps:
-                try:
-                    with open(path, "r", encoding="utf-8") as f:
-                        snap = json.load(f)
-                except Exception:
-                    continue
-                meta = snap.get("_meta", {})
-                ts = meta.get("saved_at",""); usuario = meta.get("usuario","")
-                motivo = meta.get("motivo",""); cambios = meta.get("diffs","")
-                rowtxt = (motivo + " " + cambios).lower()
-                if q and q not in rowtxt: continue
-                tree.insert("", "end", values=(ver, ts, usuario, motivo, cambios))
-
-            # --- AUTOSELECT last ---
-            items = tree.get_children()
-            if items:
-                tree.selection_set(items[-1]); tree.see(items[-1]); on_select()
-
-        def on_select(*_):
-            nonlocal last_snap
-            clear_detail()
-            sel = tree.selection()
-            if not sel:
-                detail_title.configure(text="Detalle de versión")
-                btn_xlsx.configure(state="disabled"); btn_pdf.configure(state="disabled")
-                return
-            ver = tree.item(sel[0], "values")[0]
-            snap = _load_version_snapshot(self.current_mold, ver)
-            last_snap = snap
-            meta = snap.get("_meta", {})
-            clean = dict(snap); clean.pop("_meta", None)
-
-            detail_title.configure(text=f"Detalle de versión — {ver}")
-            btn_xlsx.configure(state="normal")
-            btn_pdf.configure(state="normal")
-
-            # Parameter overview
-            g = section("Parameter overview")
-            g.grid_columnconfigure((0,1,2,3,4,5), weight=1)
-            row_labeled(g, 0, "Program",      [clean.get("program",""), "Date of entry", clean.get("date_of_entry",""),
-                                               "Cavities", clean.get("cavities","")], start_col=0)
-            row_labeled(g, 1, "Mould desig.", [clean.get("mould_desig",""), "Machine", clean.get("machine",""),
-                                               "Material", clean.get("material","")], start_col=0)
-
-            # Key data
-            k = section("Key data")
-            items = [
-                ("Cycle time [s]", "cycle_time_s"),
-                ("Injection time [s]", "injection_time_s"),
-                ("Holding press. time [s]", "holding_press_time_s"),
-                ("Rem. cooling time [s]", "rem_cooling_time_s"),
-                ("Dosage time [s]", "dosage_time_s"),
-                ("Screw stroke [mm]", "screw_stroke_mm"),
-                ("Mould stroke [mm]", "mould_stroke_mm"),
-                ("Ejector stroke [mm]", "ejector_stroke_mm"),
-                ("Shot weight [g]", "shot_weight_g"),
-                ("Plasticising flow [kg/h]", "plasticising_flow_kgh"),
-                ("Dosage capacity [g/s]", "dosage_capacity_gs"),
-                ("Dosage volume [ccm]", "dosage_volume_ccm"),
-                ("Material cushion [ccm]", "material_cushion_ccm"),
-                ("max. inj. pressure [bar]", "max_inj_pressure_bar"),
-            ]
-            k.grid_columnconfigure((0,1,2), weight=1)
-            rr = 0
-            for label, fid in items:
-                row_labeled(k, rr, label, [clean.get(fid,"")], start_col=0)
-                rr += 1
-
-            # Injection unit
-            iu = section("Injection unit")
-            iu.grid_columnconfigure((0,1,2,3), weight=1)
-            row_labeled(iu, 0, "Screw Ø [mm]", [clean.get("screw_d_mm",""), "Pcs. 1", clean.get("pcs_1","")], start_col=0)
-
-            # Injection
-            inj = section("Injection")
-            inj.grid_columnconfigure((0,1,2,3,4), weight=1)
-            row_labeled(inj, 0, "Injection press. limiting [bar]",
-                        [clean.get("inj_press_lim_1",""), clean.get("inj_press_lim_2",""), clean.get("inj_press_lim_3","")])
-            row_labeled(inj, 1, "Injection speed [mm/s]",
-                        [clean.get("inj_speed_1",""), clean.get("inj_speed_2",""), clean.get("inj_speed_3","")])
-            row_labeled(inj, 2, "End of stage [mm]",
-                        [clean.get("inj_end_stage_mm_1",""), clean.get("inj_end_stage_mm_2",""), clean.get("inj_end_stage_mm_3","")])
-            row_labeled(inj, 3, "Injection flow [ccm/s]",
-                        [clean.get("inj_flow_1",""), clean.get("inj_flow_2",""), clean.get("inj_flow_3","")])
-            row_labeled(inj, 4, "End of stage [ccm]",
-                        [clean.get("inj_end_stage_ccm_1",""), clean.get("inj_end_stage_ccm_2",""), clean.get("inj_end_stage_ccm_3","")])
-
-            # Plasticizing
-            pl = section("Plasticizing (St.1)")
-            pl.grid_columnconfigure((0,1), weight=1)
-            row_labeled(pl, 0, "Screw speed [m/min]", [clean.get("plast_screw_speed","")])
-            row_labeled(pl, 1, "Back pressure [bar]", [clean.get("plast_back_pressure","")])
-            row_labeled(pl, 2, "End of stage [ccm]", [clean.get("plast_end_stage_ccm","")])
-
-            # Holding pressure
-            hp = section("Holding pressure (Pcs.2)")
-            hp.grid_columnconfigure((0,1,2,3,4), weight=1)
-            row_labeled(hp, 0, "Time [s]",     [clean.get("hp_time_1",""), clean.get("hp_time_2",""), clean.get("hp_time_3","")])
-            row_labeled(hp, 1, "Pressure [bar]", [clean.get("hp_press_1",""), clean.get("hp_press_2",""),
-                                                  clean.get("hp_press_3",""), clean.get("hp_press_4","")])
-
-            # Temperatures
-            tp = section("Temperatures (1..5)")
-            tp.grid_columnconfigure((0,1,2,3,4,5), weight=1)
-            row_labeled(tp, 0, "Cylinder temp. [°C]", [clean.get("temp_c1",""), clean.get("temp_c2",""),
-                                                       clean.get("temp_c3",""), clean.get("temp_c4",""),
-                                                       clean.get("temp_c5","")])
-            row_labeled(tp, 1, "Tolerances [°C]", [clean.get("tol_c1",""), clean.get("tol_c2",""),
-                                                   clean.get("tol_c3",""), clean.get("tol_c4",""),
-                                                   clean.get("tol_c5","")])
-            row_labeled(tp, 2, "Feed yoke temperature [°C]", [clean.get("feed_yoke_temp","")])
-            row_labeled(tp, 3, "Lower enable tol. [°C]",     [clean.get("lower_enable_tol","")])
-            row_labeled(tp, 4, "Upper switch-off tol. [°C]", [clean.get("upper_switch_off_tol","")])
-
-            # Mould movements Opening
-            mo = section("Mould movements — Opening (St.1 / St.2 / St.3)")
-            mo.grid_columnconfigure((0,1,2,3), weight=1)
-            row_labeled(mo, 0, "End of stage [mm]", [clean.get("open_end_mm_1",""), clean.get("open_end_mm_2",""), clean.get("open_end_mm_3","")])
-            row_labeled(mo, 1, "Speed [mm/s]",      [clean.get("open_speed_1",""), clean.get("open_speed_2",""), clean.get("open_speed_3","")])
-            row_labeled(mo, 2, "Force [kN]",        [clean.get("open_force_1",""), clean.get("open_force_2",""), clean.get("open_force_3","")])
-
-            # Mould movements Closing
-            mc = section("Mould movements — Closing (St.1 / St.2 / St.3 / An. HD)")
-            mc.grid_columnconfigure((0,1,2,3,4), weight=1)
-            row_labeled(mc, 0, "End of stage [mm]", [clean.get("close_end_mm_1",""), clean.get("close_end_mm_2",""),
-                                                     clean.get("close_end_mm_3",""), clean.get("close_end_mm_4","")])
-            row_labeled(mc, 1, "Speed [mm/s]",      [clean.get("close_speed_1",""), clean.get("close_speed_2",""),
-                                                     clean.get("close_speed_3",""), clean.get("close_speed_4","")])
-            row_labeled(mc, 2, "Force [kN]",        [clean.get("close_force_1",""), clean.get("close_force_2",""),
-                                                     clean.get("close_force_3","")])
-
-            # Clamping
-            cl = section("Clamping")
-            cl.grid_columnconfigure((0,1), weight=1)
-            row_labeled(cl, 0, "Mould closed [kN]", [clean.get("mould_closed_kn","")])
-
-        # ---------------------- EXPORTS (integrados) ----------------------
-        def export_excel():
-            sel = tree.selection()
-            if not sel:
-                return
-            ver = tree.item(sel[0], "values")[0]
-            ver_label = f"V.{ver[1:]}" if str(ver).lower().startswith("v") else str(ver)
-
-            # 1) Elegir dónde guardar
-            default_name = f"{self.current_mold}_{ver}.xlsx"
-            path = filedialog.asksaveasfilename(
-                defaultextension=".xlsx",
-                filetypes=[("Excel", "*.xlsx")],
-                initialfile=default_name
-            )
-            if not path:
-                return
-
-            try:
-                # 2) Encontrar plantilla
-                template_path = _find_excel_template()
-                if not template_path:
-                    # Si no está, deja elegirla manualmente una sola vez
-                    messagebox.showwarning(
-                        "Plantilla no encontrada",
-                        "No encontré 'formatorecta.xlsx' en las rutas conocidas.\n"
-                        "Selecciona la plantilla (se usará esta vez):"
-                    )
-                    template_path = filedialog.askopenfilename(
-                        title="Selecciona formatorecta.xlsx",
-                        filetypes=[("Excel", "*.xlsx")]
-                    )
-                    if not template_path:
-                        return
-
-                # 3) Obtener snapshot del árbol (la versión seleccionada)
-                snap = _load_version_snapshot(self.current_mold, ver)
-                if not snap:
-                    messagebox.showwarning("Exportar Excel", "No se pudo cargar el snapshot de la versión.")
-                    return
-                # Limpia metadatos para que solo queden campos
-                snap_clean = dict(snap)
-                snap_clean.pop("_meta", None)
-
-                # 4) Exportar a plantilla usando el EXCEL_MAP
-                if sys.platform.startswith("win") and HAS_EXCEL_COM:
-                    # normaliza ruta destino (por si el diálogo devuelve barras '/')
-                    path = _normalize_win_path(path)
-                    _export_with_excel_com(snapshot=snap_clean, out_path=path, template_path=template_path)
-                else:
-                    # Aviso explícito: sin COM Excel podría meter reparaciones; usamos fallback seguro
-                    messagebox.showinfo(
-                        "Exportar Excel",
-                        "No se encontró Excel/pywin32. Usaré el modo 'limpio' sin vínculos ni dibujos para evitar reparaciones."
-                    )
-                    _export_snapshot_to_template(snapshot=snap_clean, out_path=path, template_path=template_path)
-
-                messagebox.showinfo("Exportar Excel", f"Archivo generado correctamente:\n{path}")
-            except ModuleNotFoundError:
-                messagebox.showwarning("Dependencia faltante",
-                    "Falta openpyxl para exportar en plantilla:\n\npip install openpyxl")
-            except Exception as e:
-                messagebox.showerror("Exportar Excel", f"No se pudo exportar:\n{e}")
-
-        def export_pdf():
-            sel = tree.selection()
-            if not sel:
-                return
-            ver = tree.item(sel[0], "values")[0]
-            ver_label = f"V.{ver[1:]}" if str(ver).lower().startswith("v") else str(ver)
-            title_txt = f"Receta {ver_label}"
-            path = filedialog.asksaveasfilename(defaultextension=".pdf",
-                                                filetypes=[("PDF", "*.pdf")],
-                                                initialfile=f"{self.current_mold}_{ver}.pdf")
-            if not path:
-                return
-            try:
-                from reportlab.lib.pagesizes import A4, landscape
-                from reportlab.pdfgen import canvas
-                from reportlab.lib.units import mm
-                from reportlab.lib.utils import ImageReader
-
-                snap = dict(last_snap); _ = snap.pop("_meta", {})
-                W, H = landscape(A4)
-                margin = 12*mm
-                col_gap = 18*mm
-                col_w = (W - margin*2 - col_gap) / 2.0
-                row_h = 10*mm
-
-                logo_path = LOGO_PATH if os.path.exists(LOGO_PATH) else ""
-                c = canvas.Canvas(path, pagesize=landscape(A4))
-                c.setStrokeColorRGB(0.7, 0.7, 0.7)
-                c.setLineWidth(0.5)
-
-                def new_page():
-                    c.showPage()
-                    return W, H, margin, H - margin
-
-                def draw_head(txt):
-                    c.setFillColorRGB(0.86, 0.92, 1)
-                    c.rect(x, y - row_h, col_w, row_h, stroke=0, fill=1)
-                    c.setFillColorRGB(0.11, 0.30, 0.85)
-                    c.setFont("Helvetica-Bold", 12)
-                    c.drawString(x + 2, y - row_h + 3, _safe_pdf(txt))
-                    c.setFillColorRGB(0, 0, 0)
-
-                def draw_box(lbl, val, w, h=row_h, center=False):
-                    c.setFillColorRGB(0.96, 0.96, 0.96)
-                    c.rect(x, y - h, w, h, stroke=1, fill=1)
-                    c.setFillColorRGB(0, 0, 0)
-                    c.setFont("Helvetica-Bold", 9)
-                    c.drawString(x+2, y - 12, _safe_pdf(lbl))
-                    if val is not None:
-                        c.setFont("Helvetica", 10)
-                        if center:
-                            c.drawCentredString(x + w/2, y - h/2 - 3, _safe_pdf(val))
-                        else:
-                            c.drawString(x+2, y - h + 4, _safe_pdf(val))
-
-                def advance(h=row_h+6):
-                    nonlocal y, x, col
-                    y -= h
-                    if y < margin + row_h*5:
-                        if col == 0:
-                            col = 1; x = margin + col_w + col_gap; y = H - margin - 24
-                        else:
-                            col = 0; x = margin
-                            _, _, _, y0 = new_page(); y = y0 - 24
-                        # logo + título por página
-                        if logo_path:
-                            try:
-                                img = ImageReader(logo_path)
-                                c.drawImage(img, W-60*mm, H-18*mm, width=50*mm, height=12*mm,
-                                            preserveAspectRatio=True, mask='auto')
-                            except Exception:
-                                pass
-                        c.setFont("Helvetica-Bold", 14)
-                        c.drawString(margin, H - margin - 8, _safe_pdf(title_txt))
-
-                # logo + título primera página
-                if logo_path:
-                    try:
-                        img = ImageReader(logo_path)
-                        c.drawImage(img, W-60*mm, H-18*mm, width=50*mm, height=12*mm,
-                                    preserveAspectRatio=True, mask='auto')
-                    except Exception:
-                        pass
-                c.setFont("Helvetica-Bold", 14)
-                c.drawString(margin, H - margin - 8, _safe_pdf(title_txt))
-
-                x, y = margin, H - margin - 24
-                col = 0
-
-                # Cabecera
-                draw_box("Program", snap.get("program",""), w=40*mm); advance()
-                draw_box("Date of entry:", snap.get("date_of_entry",""), w=60*mm); advance()
-                draw_box("Cavities", snap.get("cavities",""), w=30*mm); advance()
-                draw_box("Mould desig.", snap.get("mould_desig",""), w=40*mm); advance()
-                draw_box("Machine", snap.get("machine",""), w=40*mm); advance()
-                draw_box("Material", snap.get("material",""), w=40*mm); advance(h=10)
-
-                # Injection unit
-                draw_head("Injection unit"); advance(h=12)
-                draw_box("Screw Ø [mm]", snap.get("screw_d_mm",""), w=35*mm); advance()
-                draw_box("Pcs. 1", snap.get("pcs_1",""), w=25*mm); advance(h=14)
-
-                # Forzar salto para Key data si falta espacio
-                if y < margin + row_h*18 and col == 0:
-                    col = 1; x = margin + col_w + col_gap; y = H - margin - 24
-
-                # Key data
-                draw_head("Key data"); advance(h=12)
-                kd = [
-                    ("Cycle time [s]", "cycle_time_s"),
-                    ("Injection time [s]", "injection_time_s"),
-                    ("Holding press. time [s]", "holding_press_time_s"),
-                    ("Rem. cooling time [s]", "rem_cooling_time_s"),
-                    ("Dosage time [s]", "dosage_time_s"),
-                    ("Screw stroke [mm]", "screw_stroke_mm"),
-                    ("Mould stroke [mm]", "mould_stroke_mm"),
-                    ("Ejector stroke [mm]", "ejector_stroke_mm"),
-                    ("Shot weight [g]", "shot_weight_g"),
-                    ("Plasticising flow [kg/h]", "plasticising_flow_kgh"),
-                    ("Dosage capacity [g/s]", "dosage_capacity_gs"),
-                    ("Dosage volume [ccm]", "dosage_volume_ccm"),
-                    ("Material cushion [ccm]", "material_cushion_ccm"),
-                    ("max. inj. pressure [bar]", "max_inj_pressure_bar"),
-                ]
-                for lab, fid in kd:
-                    draw_box(lab, snap.get(fid,""), w=col_w); advance()
-
-                # Volver a primera col si conviene
-                col = 0; x = margin; y = y if y > H/2 else H - margin - 24
-
-                # Injection (3 columnas)
-                draw_head("Injection"); advance(h=12)
-                def triple(lbl, k1, k2, k3):
-                    wlbl = 65*mm; wcell = (col_w - wlbl)/3.0
-                    draw_box(lbl, None, w=wlbl)
-                    c.rect(x+wlbl, y - row_h, wcell, row_h, 1, 0)
-                    c.drawCentredString(x+wlbl + wcell/2, y - row_h/2 - 3, _safe_pdf(snap.get(k1,"")))
-                    c.rect(x+wlbl+wcell, y - row_h, wcell, row_h, 1, 0)
-                    c.drawCentredString(x+wlbl + wcell*1.5, y - row_h/2 - 3, _safe_pdf(snap.get(k2,"")))
-                    c.rect(x+wlbl+wcell*2, y - row_h, wcell, row_h, 1, 0)
-                    c.drawCentredString(x+wlbl + wcell*2.5, y - row_h/2 - 3, _safe_pdf(snap.get(k3,"")))
-                    advance()
-                triple("Injection press. limiting [bar]", "inj_press_lim_1","inj_press_lim_2","inj_press_lim_3")
-                triple("Injection speed [mm/s]",          "inj_speed_1","inj_speed_2","inj_speed_3")
-                triple("End of stage [mm]",               "inj_end_stage_mm_1","inj_end_stage_mm_2","inj_end_stage_mm_3")
-                triple("Injection flow [ccm/s]",          "inj_flow_1","inj_flow_2","inj_flow_3")
-                triple("End of stage [ccm]",              "inj_end_stage_ccm_1","inj_end_stage_ccm_2","inj_end_stage_ccm_3")
-
-                # Plasticizing
-                draw_head("Plasticizing (St.1)"); advance(h=12)
-                draw_box("Screw speed [m/min]", snap.get("plast_screw_speed",""), w=col_w); advance()
-                draw_box("Back pressure [bar]", snap.get("plast_back_pressure",""), w=col_w); advance()
-                draw_box("End of stage [ccm]",  snap.get("plast_end_stage_ccm",""), w=col_w); advance(h=10)
-
-                # Holding pressure
-                draw_head("Holding pressure (Pcs.2)"); advance(h=12)
-                triple("Time [s]", "hp_time_1","hp_time_2","hp_time_3")
-                # presión (4 columnas)
-                wlbl = 65*mm; wcell = (col_w - wlbl)/4.0
-                draw_box("Pressure [bar]", None, w=wlbl)
-                keys = ["hp_press_1","hp_press_2","hp_press_3","hp_press_4"]
-                for i,k in enumerate(keys):
-                    cx = x + wlbl + i*wcell
-                    c.rect(cx, y - row_h, wcell, row_h, 1, 0)
-                    c.drawCentredString(cx + wcell/2, y - row_h/2 - 3, _safe_pdf(snap.get(k,"")))
-                advance(h=row_h+4)
-
-                # Temperatures
-                draw_head("Temperatures"); advance(h=12)
-                def five(lbl, keys):
-                    wlbl = 65*mm; wcell = (col_w - wlbl)/5.0
-                    draw_box(lbl, None, w=wlbl)
-                    for i,k in enumerate(keys):
-                        cx = x + wlbl + i*wcell
-                        c.rect(cx, y - row_h, wcell, row_h, 1, 0)
-                        c.drawCentredString(cx + wcell/2, y - row_h/2 - 3, _safe_pdf(snap.get(k,"")))
-                    advance()
-                five("Cylinder temp. [°C]", ["temp_c1","temp_c2","temp_c3","temp_c4","temp_c5"])
-                five("Tolerances [°C]",     ["tol_c1","tol_c2","tol_c3","tol_c4","tol_c5"])
-                draw_box("Feed yoke temperature [°C]", snap.get("feed_yoke_temp",""), w=col_w); advance()
-                # doble columna en fila
-                whalf = (col_w - 4*mm)/2
-                draw_box("Lower enable tol. [°C]", snap.get("lower_enable_tol",""), w=whalf)
-                c.rect(x+whalf+4*mm, y - row_h, whalf, row_h, 1, 0)
-                c.setFont("Helvetica-Bold", 9); c.drawString(x+whalf+4*mm+2, y - 12, _safe_pdf("Upper switch-off tol. [°C]"))
-                c.setFont("Helvetica", 10); c.drawString(x+whalf+4*mm+2, y - row_h + 4, _safe_pdf(snap.get("upper_switch_off_tol","")))
-                advance(h=row_h+6)
-
-                # Opening
-                draw_head("Mould movements — Opening (St.1 / St.2 / St.3)"); advance(h=12)
-                triple("End of stage [mm]", "open_end_mm_1","open_end_mm_2","open_end_mm_3")
-                triple("Speed [mm/s]",      "open_speed_1","open_speed_2","open_speed_3")
-                triple("Force [kN]",        "open_force_1","open_force_2","open_force_3")
-
-                # Closing
-                draw_head("Mould movements — Closing (St.1 / St.2 / St.3 / An. HD)"); advance(h=12)
-                def quad(lbl, keys):
-                    wlbl = 65*mm; wcell = (col_w - wlbl)/4.0
-                    draw_box(lbl, None, w=wlbl)
-                    for i,k in enumerate(keys):
-                        cx = x + wlbl + i*wcell
-                        c.rect(cx, y - row_h, wcell, row_h, 1, 0)
-                        c.drawCentredString(cx + wcell/2, y - row_h/2 - 3, _safe_pdf(snap.get(k,"")))
-                    advance()
-                quad("End of stage [mm]", ["close_end_mm_1","close_end_mm_2","close_end_mm_3","close_end_mm_4"])
-                quad("Speed [mm/s]",      ["close_speed_1","close_speed_2","close_speed_3","close_speed_4"])
-                # 3 columnas para force
-                wlbl = 65*mm; wcell = (col_w - wlbl)/4.0
-                draw_box("Force [kN]", None, w=wlbl)
-                for i,k in enumerate(["close_force_1","close_force_2","close_force_3"]):
-                    cx = x + wlbl + i*wcell
-                    c.rect(cx, y - row_h, wcell, row_h, 1, 0)
-                    c.drawCentredString(cx + wcell/2, y - row_h/2 - 3, _safe_pdf(snap.get(k,"")))
-                advance()
-
-                # Clamping
-                draw_head("Clamping"); advance(h=12)
-                draw_box("Mould closed [kN]", snap.get("mould_closed_kn",""), w=col_w); advance()
-
-                c.showPage(); c.save()
-                messagebox.showinfo("Exportar PDF","Archivo PDF generado correctamente.")
-            except ModuleNotFoundError:
-                messagebox.showwarning("Dependencia faltante",
-                    "Para exportar PDF instala reportlab:\n\npip install reportlab")
-            except Exception as e:
-                messagebox.showerror("Exportar PDF", f"No se pudo exportar:\n{e}")
-
-        tree.bind("<<TreeviewSelect>>", on_select)
-        btn_xlsx.configure(command=export_excel)
-        btn_pdf.configure(command=export_pdf)
-        load_versions()
-
-    # --------------------------- Importar Excel simple ---------------------------
-    def _import_excel(self):
-        try:
-            from openpyxl import load_workbook
-        except Exception:
-            messagebox.showwarning("Falta dependencia", "Instala openpyxl: pip install openpyxl")
-            return
-        path = filedialog.askopenfilename(filetypes=[("Excel", "*.xlsx;*.xlsm;*.xls")])
-        if not path: return
-        try:
-            wb = load_workbook(path, data_only=True, keep_vba=True)
-            ws = wb.active
-            hits, misses = 0, 0
-            for excel_key, spec in EXCEL_MAP.items():
-                ui_key = ALIAS_EXCEL_TO_UI.get(excel_key, excel_key)
-                try:
-                    addr = _anchor_address(ws, spec)
-                    val = ws[addr].value
-                except Exception:
-                    val = None
-                if ui_key in self.vars and val is not None:
-                    self.vars[ui_key].set(_cast_numeric(val) if ui_key in NUM_FIELDS else str(val))
-                    hits += 1
-                else:
-                    misses += 1
-            self._dirty = True
-            messagebox.showinfo("Importar",
-                                f"Se mapearon {hits} campo(s) desde el Excel.\n"
-                                f"Sin valor o no presente en UI: {misses}.")
-        except Exception as e:
-            messagebox.showerror("Importar Excel", f"No se pudo leer el archivo:\n{e}")
-
-    # --------------------------- Utilidades ---------------------------
-    def _open_folder(self, path):
-        try:
-            if os.path.isdir(path):
-                if sys.platform.startswith("win"):
-                    os.startfile(path)
-                elif sys.platform == "darwin":
-                    os.system(f'open "{path}"')
-                else:
-                    os.system(f'xdg-open "{path}"')
-            else:
-                folder = os.path.dirname(path)
-                if sys.platform.startswith("win"):
-                    os.startfile(folder)
-                elif sys.platform == "darwin":
-                    os.system(f'open "{folder}"')
-                else:
-                    os.system(f'xdg-open "{folder}"')
-        except Exception as e:
-            messagebox.showwarning("Abrir carpeta", f"No se pudo abrir:\n{e}")
